% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbcc_cols.R
\name{bbcc_cols}
\alias{bbcc_cols}
\title{Use official BBCC colors}
\usage{
bbcc_cols(...)
}
\arguments{
\item{...}{}
}
\value{
A vector of colors used for BBCC branding. Also can assign color within
        ggplot
}
\description{
Use official BBCC colors
}
\examples{

mtcars \%>\%
      ggplot(aes(mpg, cyl)) +
      geom_point(color = bbcc_cols("dark blue")) +
      labs(
           title = "Some title") +
      theme(
           plot.title = element_text(size = 20, color = bbcc_cols("green"))
           )

}
