% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/program_area_fte.R
\name{program_area_ftes}
\alias{program_area_ftes}
\title{FTE by Program Area}
\usage{
program_area_ftes()
}
\description{
Function will return a dataframe containing the FTE for each academic year broken out by program area.
}
\examples{

program_area_ftes() \%>\%
  filter(year >= 'B23')

}
