% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention_rate.R
\name{retention_race}
\alias{retention_race}
\title{Calculating retention rates}
\usage{
retention_race(start_yrq, nxt_yrq, desired_qtr)
}
\arguments{
\item{start_yrq}{Indicate the cohort year such as B23, B34 and so forth}

\item{nxt_yrq}{Indicate one year out from the cohort year such as B23 will be B34 and so forth.}

\item{desired_qtr}{Indicated which quarter to quarter retention rate you would like to look at.}
}
\description{
Calculating retention rates
}
