% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sai_points_calc.R
\name{sai_calculation}
\alias{sai_calculation}
\title{This function will produce a}
\usage{
sai_calculation(sai_area, sai_title)
}
\arguments{
\item{sai_area}{This is the SAI point that is going to be calcuated whic may include 1st 15 credits, completion, retention or some
point. The ABE point is not calculated in this manner}

\item{sai_title}{String that will be used for the title of the table that is generated used the `sai_tbl` function.}
}
\value{
A tibble containing a ggplot object, kable table, and relevant data that can be plucked and used in the report.
}
\description{
This function will produce a
}
\examples{




}
