% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bbcc.R
\name{theme_bbcc}
\alias{theme_bbcc}
\title{BBCC Theme}
\usage{
theme_bbcc()
}
<<<<<<< HEAD
\value{
Thematic settings for ggplot object
}
\description{
A theme that will customize graphics according standards.
}
\examples{

mtcars \%>\%
   ggplot(aes(mpg, cyl)) +
   geom_point() +
   labs(
        title = "Some title") +
   theme_bbcc() +
   theme(
        plot.title = element_text(size = 16)
        )


=======
\description{
A theme that will customize graphics according standards.
>>>>>>> 53f0d1347ec3751a96cc4109a042091d2c2fcc8c
}
