% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modality_fte.R
\name{modality_fte}
\alias{modality_fte}
\title{Modality Enrollment}
\usage{
modality_fte()
}
\description{
Get a dataframe of fte for each modality since the start.
}
\examples{

 modality_fte() \%>\%
   filter(year_long == "2018-19")

}
